worked on more tests
